Reverting a change to the include path
Sorry for reverting that, but I got many e-mails it said it was not compoling anymore. It was not compoling for me.
It also broke the official and automated build on launchpad and OBS, so I revert this. It is -not- a solution,
but it seem to be better for a vast majority of users. This is a temporary change so I stop getting emails from users
and bots about it.