[ #16349 ] Refector the canvas tip so the 'loader' tasks delegate to subclasses of 'tip' instead of hardcoding them (1/2)
From now on, the loader class is dedicated to painting the tip on the canvas. The image loading is delegated to the Tip itself.
In the future, the tip will stop using SVG and have it's own painter and the 'loader' class will be renamed something else because
it wont load much anymore. This is required to implement multiple tips, animations and cpu-friendly compositing.