[ #41915 ] Add partial support for a new backend managing abstraction
This commit add the ability to work with item backends as an abstract group
of items (calls, contacts, bookmark, ...). Backends can be enabled, disabled
and eventually maipulated as a group. Use cases for this include features
such as:

 * Ability to track presense on an LDAP group
 * Default incomming call action (auto answer, DnD, forward)
 * Add prefix on outgoing call
 * Disable and enable data source
 * Rename and change icon
 * Add new data sources as script such as gravatar or YellowPages

This commit make it load, it isn't useful yet, but this is a big change,
so better merge it now and slowly implement the missing calls.