[ #49034 ] Refactor the video configuration code to be safer (1/3)
This commit implement the new daemon video getcapabilities() to load
the whole device configuration when using the video module for the
first time. This prevent the back and forth between the various
layers of both the client and the daemon. Those, created a feedback
loop that could crash the daemon or corrupt the Qt models.

While it will take a while to get to a stable point again, this
code is much more robust than the previous implementation.

The next few commits will replace the last bits of the old code
path by the new one.