[ #51339 ] Remove broken features from the account wizard
It is now possible to also disable the account wizard. This
is useful for IP2IP or DNS based registrar users.

Sflphone.org free registrar sill exist, but is unreliable and
unsupported by our team. This part of the wizard never worked
well and caused many lock-ups. Not something you want to show
new users.

The account wizard need to be rewritten, this have been planned
for years but is never a top priority. This step will at least
avoid most errors.