video: Always copy the frame
This will cost a lot of CPU, but the daemon regressions caused the
client to crash over 10 times since the video was introduced.

I no longer trust the daemon to keep its memory contracts and will
copy the frame. Given the maintainance this peice of code require,
I am not willing to do full QA cycles for every since daemon.

At the cost of additional CPU, this commit ensure video stability.