text: Implement out of call text messaging.
*This relies on yet to be merged Ring-LRC commits.*

This commit adds the ability to display the text chat on incoming
text messaging without an active call. This was a new feature in
Alpha 1, but was never properly implemented by the KDE client.

As ring matures, this become an increasignly important use case,
if not the most important of them all. There was previously some
hopes that the text messages could be handled using Telepathy-KDE,
but that effort has stalled, so here's some native support.