cleanup: Delete the macro subsystem
This isn't very useful for the vast majority of users and it
almost never tested.

The main reason to have this was botting through calls menu and
adding shortcuts to specific phone number. The second use case
is more relevant, but the macro implementation was unsuited to
make it easier to configure.