cleanup: Remove the hook subsystem
Honestly, I am not even sure what's the use case for this. Someone
once paid for it and had a weird custom SIP server. I guess if
someone really, really want to do this, the Python API is the
place to do it.

It allowed (if it worked, I am not even capable of testing it) to
run commands and open apps on some MIME payload in the second part
of SIP packets.