cleanup: Hide many options for Ring accounts.
The SIP compliance tests and mandatory options are very painful
and verbose. That's a fact of life and compliance is important.
That being said, Ring accounts don't have to endure such madness.
Many options can be safely hidden and only available in the
config file as they are required for < 0.05% of users. This
mostly revolve around bypassing strict Firewalls. The following
options are now hidden:

 * Audio port range: Required when hardcoding firewall rules for VoIP media
   instead of using STUN/TURN/UPnP/NatPNP/ICE. For Ring accounts, this
   would "work", but it is very confusing and not used often enough to be
   relevant.

 * DTMF type: This is mandatory to be compliant with the RFC, however RING
   accounts rarely use DTMF for bots like classical phone menus and even
   if they did, it would probably only support the default method.

 * Published address: Mandatory to spoof some firewall or servers to
   think Ring-KDE is a valid enterprise client. If Ring protocol users
   face such misconfigured Firewall, then other things will break too.