lrc: Use the master branch for the Ring-LRC fork.
This project no longer use the upstream ring-lrc. It is now mostly
unmaintained and the 2 maintainers that took over after me left
the project. Given the size and velocity of the changes, trying
to upstream the patches was no longer viable. The project will
soon be merged back into this repository (where it started).