ringapp: Partially revert the commit that introduced the singleton.
It moves back some of the logic into main.cpp. This is due to:

    https://bugreports.qt.io/browse/QTBUG-52988

It does **not** solve the invalid read, the seem to reduce the
likelyhood of the segmentation fault by about 50%, as mentionned
in the bug report. Valgrind still reports the invalid read and it
sometime (ok, rather often) still crash, but not always.