timelineWindow: Avoid a double free.
When the window is closed by X11 instead of the KStandardAction, it
would have been freed twice (and crash).

The reason why is *has* to be destroyed in the QApplication instead
of letting the QObject tree do its job is because of a QML bug where
the cache can hold QSharedPointer for too long (even if they eventually
GC). This prevent Ring-KDE from quitting as the existing shared
pointer cause code to be executed after the object has been deleted.

But the original implementation didn't take into account some basic
X11 wisdom not to trust anything. The MainWindow was also the parent
QObject of some objects that should have used the
QCoreApplication::instance() as parent, but this was fixed last week.