Fix (almost all) translations for QML files
KDeclarative is already used, so switch to i18n for .qml files
as well.
Fix Messages.sh to extract the strings from those files as well.

There is one caveat: src/callview/qml/videocontroltoolbar.qml
was not fixed, because when qsTr calls are replaced by i18n,
the program crashes when the video configuration panel is open.
It requires more investigation, but for now let's increase
the number of strings that can be translated.