cmake: Remove KIO and KCompletion dependencies.
They made developing the mobile and appimage version too complex.

As the app is being rewritten in QML, they will have to go. The
sooner the better.

This commit remove some of the old UI features instead of porting
them. Nothing important was lost. Most features are already
available in the new UI.