timeline: Mitigate a resizing race condition
The proper fix is to delay the QModelIndex QQuickItem in
the FlickableView and QuickTreeView classes. However after
a couple hours, it seems it's just easier to bypass the issue
with a simple timer until lazy loading is fully implemented.
All the fix that can currently be done will just make implementing
lazy loading harder so this hack is better for now.