Expose better error detection to the GUI (1/3)
This commit moves the header and timeline to the new Individual
object introduced a while back to help solve these kind of issues.

There is about a dozen of corner cases that were previously unhandled.

However there's much more to come, this commit only pushes what's
mostly stable. It breaks the statistics in the Contact tab, but this
isn't a very important issue for now.