treeview: Split the up/down/left/right code between the 2 layers.
Previously, the "visual" layer was the only one to implement the
directional iterators. However now that collapsed items or items
that fail to load are [to be] supported, this is no longer a 1-1
mapping.

This commit refactor the existing code to allow the 2 representations
to coexist independently. The visual representation is a superset
of the model topology.