abstractview: Split the model tracking code from the view.
The next commit will move it to a new file. This will prevent
AbstractQuickView from becoming a god class. Also, it will eventually
allow to share the model tracking for multiple views. A bit like the
old QGraphicsView scenes view concept. Such optimization would be
premature for now, but for charts and graphs might eventually be
relevant.