abstractview: Get rid of the `ModelIndexItem` class.
Until now, the selection handling was in FlickableView, which is below
the normal item view class in the inheritance stack. It will now be
moved above it, so it wont be necessary anymore to keep this interface.

The next step is to merge the public API into AbstractViewItem and
have a private API for the TreeTraversalReflector.