wizard: Remove the spacing for the handle.
A Kirigami commit recently introduced the ability to prevent the handle
from ever getting to the bottom right. It is also now fully hidden
during the wizard anyway (which now work properly on mobile).

The extra spacing the make sure the back button wasn't covered is no
longer required.

BUG 404023